Fixes an rpmsg bug concerning available buffers
The vring implementation used by rpmsg contains a free
running index for the last available buffer that was
filled by the ARM side. The PRU rpmsg library also keeps
track locally of the last available buffer index that it
consumed in the form of a free running index.

The PRU rpmsg library can compare the last available index
that it consumed with the last available index reported by
the ARM core (to the vring) to determine if there is a new
buffer to be consumed. If the two available indexes are
equal to one another, then the PRU rpmsg library can tell
that there are no messages waiting for it in the vring.

Both of these indexes need to be of the same type in order
for unsigned integer overflow to affect them equally. This
was not the case prior to this commit. Previously, the
vring index was 16 bits and the local rpmsg library index
was 32 bits. Once the vring index rolled over from 0xffff
back to 0, the PRU continuously thought that a new message
was available because of the index mismatch. In the echo
example case this caused the PRU to send 65536 unsolicited
messages to the ARM, bringing it to a crawl due to all of
interrupts associated with the messages.

This commit corrects the size of the local rpmsg library's
last_avail_idx to use 16 bits instead of 32.

+++ hc-sr04-range-sensor/include/pru_virtqueue.h
@@ -94,7 +94,7 @@ struct pru_virtqueue {
 	uint32_t	id;
 	uint32_t	to_arm_event;
 	uint32_t	from_arm_event;
-	uint32_t 	last_avail_idx;
+	uint16_t 	last_avail_idx;
 	struct vring 	vring;
 };
